add icon to domain lists and dashboard
