+++ src/components/EntityIcon/index.ts
@@ -1,1 +1,1 @@
-export { default } from './EntityIcon';
+export { default, getStatusForDomain } from './EntityIcon';
