+++ src/components/StatusIndicator/StatusIndicator.tsx
@@ -62,18 +62,3 @@ const StatusIndicator = (props: Props & WithStyles<CSSClasses>) => {
 const styled = withStyles(styles);
 
 export default styled(StatusIndicator);
-
-export const getStatusForDomain = (status: string) => {
-  switch (status) {
-    case 'edit_mode':
-      return 'loading';
-    case 'active':
-      return 'active';
-    case 'disabled':
-      return 'inactive';
-    case 'has_errors':
-      return 'error';
-    default:
-      return 'inactive';
-  }
-};
