+++ src/components/StatusIndicator/index.ts
@@ -1,5 +1,2 @@
 import { Props as _Props } from './StatusIndicator';
-
-export { default, getStatusForDomain } from './StatusIndicator';
-
 export interface Props extends _Props {}
