+++ src/features/Dashboard/DomainsDashboardCard/DomainsDashboardCard.tsx
@@ -2,7 +2,6 @@ import { compose, prop, sortBy, take } from 'ramda';
 import * as React from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import DomainIcon from 'src/assets/addnewmenu/domain.svg';
 import Paper from 'src/components/core/Paper';
 import {
   StyleRulesCallback,
@@ -13,10 +12,8 @@ import Table from 'src/components/core/Table';
 import TableBody from 'src/components/core/TableBody';
 import TableCell from 'src/components/core/TableCell';
 import Typography from 'src/components/core/Typography';
+import EntityIcon, { getStatusForDomain } from 'src/components/EntityIcon';
 import Grid from 'src/components/Grid';
-import StatusIndicator, {
-  getStatusForDomain
-} from 'src/components/StatusIndicator';
 import TableRow from 'src/components/TableRow';
 import TableRowEmptyState from 'src/components/TableRowEmptyState';
 import TableRowError from 'src/components/TableRowError';
@@ -30,29 +27,15 @@ import DashboardCard from '../DashboardCard';
 
 type ClassNames =
   | 'root'
-  | 'icon'
   | 'labelGridWrapper'
   | 'description'
   | 'labelStatusWrapper'
-  | 'statusOuter'
   | 'labelCol'
   | 'actionsCol'
   | 'wrapHeader';
 
 const styles: StyleRulesCallback<ClassNames> = theme => ({
   root: {},
-  icon: {
-    position: 'relative',
-    top: 3,
-    width: 40,
-    height: 40,
-    '& .circle': {
-      fill: theme.bg.offWhiteDT
-    },
-    '& .outerCircle': {
-      stroke: theme.bg.main
-    }
-  },
   labelGridWrapper: {
     paddingLeft: '4px !important',
     paddingRight: '4px !important'
@@ -64,12 +47,6 @@ const styles: StyleRulesCallback<ClassNames> = theme => ({
     display: 'flex',
     flexFlow: 'row nowrap',
     alignItems: 'center'
-  },
-  statusOuter: {
-    top: 0,
-    position: 'relative',
-    marginLeft: 4,
-    lineHeight: '0.8rem'
   },
   labelCol: {
     width: '90%'
@@ -145,16 +122,18 @@ class DomainsDashboardCard extends React.Component<CombinedProps, State> {
           <Link to={`/domains/${id}/records`} className={'black nu block'}>
             <Grid container wrap="nowrap" alignItems="center">
               <Grid item className="py0">
-                <DomainIcon className={classes.icon} />
+                <EntityIcon
+                  variant="domain"
+                  status={getStatusForDomain(status)}
+                  marginTop={3}
+                  loading={getStatusForDomain(status) === 'edit'}
+                />
               </Grid>
               <Grid item className={classes.labelGridWrapper}>
                 <div className={classes.labelStatusWrapper}>
                   <Typography role="header" variant="h3" data-qa-label>
                     {domain}
                   </Typography>
-                  <div className={classes.statusOuter}>
-                    <StatusIndicator status={getStatusForDomain(status)} />
-                  </div>
                 </div>
                 <Typography className={classes.description}>{type}</Typography>
               </Grid>
