+++ src/features/Domains/DomainTableRow.tsx
@@ -1,28 +1,20 @@
 import * as React from 'react';
 import { Link } from 'react-router-dom';
-import DomainIcon from 'src/assets/addnewmenu/domain.svg';
 import {
   StyleRulesCallback,
   WithStyles,
   withStyles
 } from 'src/components/core/styles';
 import Typography from 'src/components/core/Typography';
+import EntityIcon, { getStatusForDomain } from 'src/components/EntityIcon';
 import Grid from 'src/components/Grid';
-import StatusIndicator, {
-  getStatusForDomain
-} from 'src/components/StatusIndicator';
+import {} from 'src/components/StatusIndicator';
 import TableCell from 'src/components/TableCell';
 import TableRow from 'src/components/TableRow';
 import Tags from 'src/components/Tags';
 import ActionMenu from './DomainActionMenu';
 
-type ClassNames =
-  | 'domain'
-  | 'icon'
-  | 'labelStatusWrapper'
-  | 'statusOuter'
-  | 'tagWrapper'
-  | 'domainRow';
+type ClassNames = 'domain' | 'labelStatusWrapper' | 'tagWrapper' | 'domainRow';
 
 const styles: StyleRulesCallback<ClassNames> = theme => ({
   domain: {
@@ -31,28 +23,10 @@ const styles: StyleRulesCallback<ClassNames> = theme => ({
   domainRow: {
     backgroundColor: theme.bg.white
   },
-  icon: {
-    position: 'relative',
-    top: 3,
-    width: 40,
-    height: 40,
-    '& .circle': {
-      fill: theme.bg.offWhiteDT
-    },
-    '& .outerCircle': {
-      stroke: theme.bg.main
-    }
-  },
   labelStatusWrapper: {
     display: 'flex',
     flexFlow: 'row nowrap',
     alignItems: 'center'
-  },
-  statusOuter: {
-    top: 0,
-    position: 'relative',
-    marginLeft: 4,
-    lineHeight: '0.8rem'
   },
   tagWrapper: {
     marginTop: theme.spacing.unit / 2,
@@ -88,16 +62,18 @@ const DomainTableRow: React.StatelessComponent<CombinedProps> = props => {
         <Link to={`/domains/${id}`}>
           <Grid container wrap="nowrap" alignItems="center">
             <Grid item className="py0">
-              <DomainIcon className={classes.icon} />
+              <EntityIcon
+                variant="domain"
+                status={getStatusForDomain(status)}
+                marginTop={3}
+                loading={getStatusForDomain(status) === 'edit'}
+              />
             </Grid>
             <Grid item>
               <div className={classes.labelStatusWrapper}>
                 <Typography role="header" variant="h3" data-qa-label>
                   {domain}
                 </Typography>
-                <div className={classes.statusOuter}>
-                  <StatusIndicator status={getStatusForDomain(status)} />
-                </div>
               </div>
               <div className={classes.tagWrapper}>
                 <Tags tags={tags} />
